Migrando os teste de FrontControllerTest para suas respectivas ControllerTest, CartoesControllerTest, ClientesControllerTest, ContasControllerTest e FaturasControllerTest e instanciando novos objetos dessas classes na Main
