+++ src/test/java/org/example/controller/ClientesControllerTest.java
@@ -1,0 +1,74 @@
+package org.example.controller;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class ClientesControllerTest {
+    private ClientesController controller;
+
+    @BeforeEach
+    public void setup() {
+        controller = new ClientesController();
+    }
+
+    @Test
+    public void quandoComandoEhClientesEntaoExibaOpcoesDeClientes() {
+        var resultadoEsperado = """
+                ------------------------------
+                | atualizar {cpf do cliente} |
+                | cadastrar                  |
+                | desativar {cpf do cliente} |
+                | pesquisar                  |
+                ------------------------------""";
+        var resultadoReal = controller.executar("clientes");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhClientesAtualizarEntaoAtualizeOsClientes() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("clientes atualizar");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhClientesCadastrarEntaoCadastreOsClientes() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("clientes cadastrar");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhClientesDesativarEntaoDesativeOsClientes() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("clientes desativar");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhClientesPesquisarEntaoExibaOpcoesDeClientesPesquisar() {
+        var resultadoEsperado = """
+                --------------------------
+                | cpf {cpf do cliente}   |
+                | nome {nome do cliente} |
+                -------------------------""";
+        var resultadoReal = controller.executar("clientes pesquisar");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhClientesPesquisarCpfEntaoExibaDetalhesDosClientes() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("clientes pesquisar cpf");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhClientesPesquisarNomeEntaoExibaDetalhesDosClientes() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("clientes pesquisar nome");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+}
