+++ src/test/java/org/example/controller/ContasControllerTest.java
@@ -1,0 +1,83 @@
+package org.example.controller;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class ContasControllerTest {
+
+    private ContasController controller;
+
+    @BeforeEach
+    public void setup() {
+        controller = new ContasController();
+    }
+
+    @Test
+    public void quandoComandoEhContasEntaoExibaOpcoesDeContas() {
+        var resultadoEsperado = """
+                -------------------------------
+                | cadastrar                   |
+                | desativar {numero da conta} |
+                | extrato {numero da conta}   |
+                | pesquisar                   |
+                -------------------------------""";
+        var resultadoReal = controller.executar("contas");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhContasCadastrarEntaoCadastreAsContas() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("contas cadastrar");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhContasDesativarEntaoDesativeAsContas() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("contas desativar");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhContasExtratoEntaoExibaOExtradoDasContas() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("contas extrato");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhContasPesquisarEntaoExibaOpcoesDeContasPesquisar() {
+        var resultadoEsperado = """
+                ----------------------------------
+                | cpf-titular {cpf do cliente}   |
+                | nome-titular {nome do cliente} |
+                | numero {número da conta}       |
+                ----------------------------------""";
+        var resultadoReal = controller.executar("contas pesquisar");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhContasPesquisarCpfTitularEntaoExibaAsContasEncontradas() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("contas pesquisar cpf-titular");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhContasPesquisarNomeTitularEntaoExibaAsContasEncontradas() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("contas pesquisar nome-titular");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+
+    @Test
+    public void quandoComandoEhContasPesquisarNumeroEntaoExibaDetalhesDaConta() {
+        var resultadoEsperado = "não implementado";
+        var resultadoReal = controller.executar("contas pesquisar numero");
+        assertEquals(resultadoEsperado, resultadoReal);
+    }
+}
